refactor: simplify logic by applying Roslyn optimizations and .NET 8 conventions
