+++ Agendai/App.axaml.cs
@@ -59,7 +59,7 @@ public partial class App : Application
 		base.OnFrameworkInitializationCompleted();
 	}
 
-	private void DisableAvaloniaDataAnnotationValidation()
+	private static void DisableAvaloniaDataAnnotationValidation()
 	{
 		var dataValidationPluginsToRemove =
 				BindingPlugins.DataValidators.OfType<DataAnnotationsValidationPlugin>()
