+++ Agendai/ViewLocator.cs
@@ -23,7 +23,7 @@ public class ViewLocator : IDataTemplate
 			HomeWindowViewModel => new HomeWindow(),
 			AgendaWindowViewModel => new AgendaWindow(),
 			TodoWindowViewModel => new TodoWindow(),
-			PomodoroWindowViewModel => new PomodoroWindow(),
+			PomodoroWindowViewModel => new Views.Windows.PomodoroWindow.PomodoroWindow(),
 			_ => throw new ArgumentException($"No view found for {param.GetType().Name}"),
 		};
