+++ Agendai/Views/Windows/HomeWindow/HomeWindow.axaml.cs
@@ -1,7 +1,4 @@
-using Agendai.ViewModels;
-using Avalonia;
 using Avalonia.Controls;
-using Avalonia.Markup.Xaml;
 
 
 namespace Agendai.Views.Windows.HomeWindow;
