+++ Agendai/Views/Windows/PomodoroWindow/PomodoroWindow.axaml
@@ -1,5 +1,5 @@
 <UserControl
-    x:Class="Agendai.PomodoroWindow"
+    x:Class="Agendai.Views.Windows.PomodoroWindow.PomodoroWindow"
     xmlns="https://github.com/avaloniaui"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
