+++ Agendai/Views/Windows/TodoWindow/TodoWindow.axaml.cs
@@ -1,4 +1,3 @@
-using Agendai.ViewModels;
 using Avalonia.Controls;
